[MOV] mail: move forward and reply-all actions to the web
Since odoo/odoo#186875, forward and reply-all actions have been added to the
message actions. These actions use the full composer and the full composer is a
web feature (at least for now). This PR moves the related files to
`mail/static/src/core/web`.
