+++ addons/mail/static/src/core/common/message_actions.js
@@ -88,26 +88,6 @@ messageActionsRegistry
         onClick: (component) => component.props.message.unfollow(),
         sequence: 60,
     })
-    .add("reply-all", {
-        condition: (component) =>
-            component.props.message.canReplyAllandForward(component.props.thread),
-        icon: "fa fa-reply",
-        title: _t("Reply All"),
-        onClick: (component) => {
-            component.onClickMessageReplyAll();
-        },
-        sequence: 71,
-    })
-    .add("forward", {
-        condition: (component) =>
-            component.props.message.canReplyAllandForward(component.props.thread),
-        icon: "fa fa-share",
-        title: _t("Forward"),
-        onClick: (component) => {
-            component.onClickMessageForward();
-        },
-        sequence: 72,
-    })
     .add("edit", {
         condition: (component) => component.props.message.editable,
         icon: "fa fa-pencil",
